Adding privacy + Making DB more slim

+++ midnight.py
@@ -17,8 +17,9 @@ def send_simple_message(some_list):
     return requests.post(
         "https://api.mailgun.net/v3/" + HOST + "/messages",
         auth=("api", API_KEY),
-        data={"from": u"Hacker News - Daily Report <{0}@{1}>".format(USER, HOST),
-              "to": addresses,
+        data={"from": u'"Hacker News - Daily Report "<{0}@{1}>'.format(USER, HOST),
+              "to": '<{0}@{1}>'.format(USER, HOST),
+              "bcc": addresses,
               "subject": "Hacker News Update",
               "html": u'<html>{0}<br>To be removed from this mailing list write an email with the subject \'unsubscribe\' to hackernewsletter-request@freelists.org</html>'.format(some_list)})
 
@@ -28,11 +29,20 @@ def main():
     cur = conn.cursor()
     top_ten = database.get_top_ten(cur)
     for result in top_ten:
-        story = json.loads(result[2])
-        message += u'{0} <a href="{1}">{2}</a> (<a href="https://news.ycombinator.com/item?id={3}">Comments</a>)<br>'.format(result[1], story['url'], cgi.escape(story['title']), story['id'])
-    database.was_sent(cur, top_ten)
-    database.delete_unsent(cur)
+        # Id INTEGER PRIMARY KEY, Score INTEGER, Url VARCHAR, Title VARCHAR, Sent
+        number = result[0]
+        score = result[1]
+        url = result[2]
+        title = cgi.escape(result[3])
+        if url == '':
+            link = title
+        else:
+            link = '<a href="{0}">{1}</a>'.format(url, title)
+        message += u'{0} {1} (<a href="https://news.ycombinator.com/item?id={2}">Comments</a>)<br>'.format(score, link, number)
     result = send_simple_message(message)
+    if result.status_code == 200:
+        database.was_sent(cur, top_ten)
+        database.delete_unsent(cur)
     conn.commit()
     conn.close()
